server/market: send revoke order notes on suspend with purge
The market `'suspension'` route is a somewhat mixed purpose notification 
in that all connected clients receive a notification of scheduled future 
suspension, but only `'orderbook'` subscribers get the _sequenced_ 
suspension notification at the time the market actually clears the book.

As such, if you have orders on the book, but do *not* have an orderbook 
subscription (e.g. you closed the browser window and it unsubscribed 
after 30 seconds or so), you will not get the final suspension 
notification that performs the "auto-revokes" and sends the kinder 
`TopicOrderAutoRevoked` notes to the frontend.  Instead, your client 
thinks the orders are still booked **until you reconnect with the server 
after market resumption** and order status reconciles from booked to 
revoked.  This is a bug.

To ensure that non-subscribers do not keep their orders in booked status 
incorrectly, this modifies `(*Market).PurgeBook` to send a 
`revoke_order` notification to each (user,order) that was purged.  These 
do not require an `orderbook` subscription.  These revokes are less 
friendly looking because (a) they don't explain why it was revoked other 
than by observing the prior suspension scheduled notification, and (b) 
they are sent to the frontend at error level severity instead of warn as 
the auto-revokes do.  However, this fixes the bug of clients without an 
orderbook subscription being unaware that their orders were unbooked.

Finally, this makes the revoke_order notifications come *after* the 
trade suspension notification so that clients that did have a 
subscription will get the friendlier message on the frontend, and then 
simply ignore the explicit redundant revoke_order notes that follow.