ui: more nuanced match statuses and swap steps
This updates the match status text on the order status page to more
accurately reflect the possible pending actions for a revoked match.

This also updates the visibility toggling logic for the transaction
placeholders (swap/redeem/revoke) to account for uncertainty of the
outcome when the taker is waiting to see if the maker will redeem
their swap on the sly.