client/asset: bond creation and refund
This adds the client/asset.Bonder interface for creating, broadcasting,
and refunding fidelity bonds.

This updates the client/asset.Wallet interface with separate
ContractLockTimeExpired and LockTimeExpired methods, where the former
operates on the "contract data" such as an atomic swap redeem script,
and the latter checks a time.Time. ContractLockTimeExpired is the
existing method used by the atomic swap code.

client/asset/dcr: implement asset.Bonder for Decred

Decred is the initial Bonder asset. This implements the MakeBondTx,
RefundBond, and SendTransaction methods.