client/core: mmbot fixes
This fixes eth's assetWallet so that it now satisfies the
asset.BotWallet interface.

This also updates the new asset.Wallet methods, SingleLotSwapFees and
SingleLotRedeemFees, to use the update PreSwapForm and PreRedeemForm
structs that no longer have dex.Asset fields.

This also fixes the (*Core).feeEstimates so that the fallback calls
to SingleLotSwapFees and SingleLotRedeemFees use the correct wallet,
and reset the err value to nil on success.