client/asset/dcr: Unlock must unlock unmixed account
All accounts must be unlocked, including any unmixed account, which is
used to supply keys to the refund path of the swap contract script.
If using a mixed account configuration with individually-encrypted
accounts, AND the unmixed account is not unlocked, preparing the
backup refund tx in Swap will fail (good and early!).