core: Refactor MM Bot
Refactoring in preparation to add an arbitrage bot.

The `MakerBot` type handles all the interactions core and the DEX
markets. The logic of when to buy, sell, and cancel existing orders
is handled in a `BotEngine`, which is a property of the `MakerBot`. The
`MakerBot` sends notifications to the to a `BotEngine` using its `Notify`
function, currently only to let it know that a new epoch has begun. The
`BotEngine` exposes a `TradeFeed` which the `MakerBot` listens to in order
to know when to submit orders to the dex. Additional information, such
as the current orders the bot has on the order books, the current mid-gap
price of the market, are available to the `BotEngine` to query through
an interface that is provided when the engine is created.

Currently there is one `BotEngine` type, `GapEngine`, which is the same
functionality that currently exists in the `MakerBot`.