client: asynchronous order submission
This enables orders placed from the UI to be queued to improve UI/UX.
While all orders are processed in a gouroutine, orders placed from the UI
get a temporary order ID early and wait for an order update. It also ensures
queued orders submitted from the UI are accounted for until the order has
been validated and verified by a dex server.