client/core: do not revoke trade when the cancel is revoked
This fixes a bug where a revoke_order notification from the server that
refers to a cancel order ID would cause the client to revoke the
targeted trade order instead.

Most of the time a cancel order is declared stale and deleted after it
becomes 2 epochs old, a heuristic we use to infer when we've missed the
preimage request for the cancel order.  This was necessary since unlike
trade orders, we don't expect to get status of cancel orders.  However,
we neglected the possibility that we could miss the preimage request
but reconnect in time to receive the revoke_order notification at the
end of the epoch collection period.

When handleRevokeOrderMsg detects that the order ID is a cancel order,
we now unlink the cancel order and send a TopicFailedCancel order ntfn.
The TopicFailedCancel topic previously was only used when cancel orders
were declared "stale", meaning they were more than 2 epochs old.