client: check for zero fiat rates and response status code
This increases the request interval for fiat rates to 12min due to a
possibility of hitting source's request limits. This can be considered a
temporary solution since we need to adjust this interval for every added
asset.

This also checks the HTTP response status code instead of assuming
the response is the expected payload.