client/asset: fix estimateSwap split rejection
The enough func in estimateSwap does not account for a split transaction
at the start, so it is possible that funding for trySplit would actually
choose more UTXOs. Actual order funding accounts for this. For this
estimate, we will just not use a split tx if the split-adjusted required
funds exceeds the total value of the UTXOs selected with this enough
closure.

The fix is to reject a split transaction if the output amount plus
split tx fees is more than the sum of the amounts of the utxos selected
by fund, rather than the total of the available utxos provided to fund
initially. There is no point to calling fund again with a different
enough func that accounts for the cost split tx because this indicates
it would reqiure an additional UTXO, thus making the spit txn wasteful.