multi: CEX connection support, implement Binance
A new CEX interface is introduced which allows interaction with a CEX
API. One implementation of this interface is added for Binance.

A new field is added to the  `Core` type to store connections to CEXes.
The DB is extended to store API credentials for CEXes. New webserver
routes are added to register and update CEX credentials, and to connect
and disconnect from a CEX.