client/core,app: Create arbitrage bot
The `makerBot` is refactored to take a `botEngine` as a field
which contains the logic of when to place and cancel orders.
The old maker bot logic is placed into the `gapEngine` type,
and an arbitrage bot is added. The arbitrage bot compares the
DEX and the CEX order books, and if there is an arbitrage
opportunity, executes a trade on both.

The UI is also updated to be able to create both gap and arbitrage
bots.