client/{asset,core,webserver}: BondLocked balance category
This adds the client.WalletBalance.BondLocked field, which parallels
the ContractLocked field. These are used for core to supplement the
wallet's reported balance with values that the wallet does not
include in the balance itself because they are not spendable by the
wallet alone (core authors the spending tx).

This also updates the web UI's TypeScript to include bondLocked
when appropriate.  This also fixes a bug where contractLocked was
not included on the wallets page's "Locked" balance category.

This also adds to DCR's implementation fields for reserves tracking.
The values are included in the reported asset.Balance. Subsequent
commits will add the ability modify the reserves, and for funding
methods to respect the reserves.