client/asset/dcr: respect reserves in funding
This updates the DCR wallet's transaction funding method and helpers to
respect the reserves added in the previous commit.

To do so without creating sizing transactions to actively isolate the
reserved value, we add new coin selection functions that we use exclude
UTXOs from transaction funding. The objective of these coin selection
functions is different from order funding coin selection.

The main helper, leastOverFund, attempts to pick a subset of the provided
UTXOs to reach the required amount with the objective of minimizing the
total amount of the selected UTXOs. This is different from the objective
used when funding orders, which is to minimize the number of UTXOs (to
minimize fees).

NOTE: The provided UTXO set MUST be sorted in ascending order (smallest
first, largest last)!

leastOverFund begins by partitioning the UTXO slice before the smallest
single UTXO that is large enough to fully fund the requested amount, if
it exists. If the smaller set is insufficient, the single largest UTXO
is returned. If instead the set of smaller UTXOs has enough total
value, it will search for a subset that reaches the amount with least
over-funding (see subsetSmallBias and subsetLargeBias). If that subset
has less combined value than the single sufficiently-large UTXO (if it
exists), the subset will be returned, otherwise the single UTXO will be
returned.

We also modify the "enough" functions and their constructors with
excess (change) reporting in a new return from the enough function.
The constructor also accepts a flag indicating if this returned value
should be set to zero. The tryFund helper now returns the final excess
value when it completes UTXO selection. This is used in the
(*ExchangeWallet).fund method, and in other callers, to discern if
there would be sufficient remaining in the wallet (including this
change) to satisfy a reserved amount. This amount is an input to fund
called "keep". If change should not be considered "kept" (e.g. no
preceding split txn, or mixing sends change to umixed account where it
is unusable for reserves), caller should return 0 extra from the enough
func (via the enough func's constructor flag).