app: Fix swap fee summary (#2057)
* app: Fix swap fee summary

The fee estimates for tokens on the UI was not accounting for
the existance of tokens. When trading tokens, the fee is paid in
the parent asset, so it doesn't make sense to display the fee as a
percentage of the order amount. The fee summary pane now displays
the actual fee estimate ranges instead of the percentage, and the
fee details pane only shows percentages for non-token assets.

* Update dash and fix formatting