eth: remove isRedeemable contract method
This method made more sense when the client was using a private node,
however this is a vulnerability with public RPC providers.

This removes the isRedeemable public method from the Solidity contracts
ETHSwapV0 and ERC20SwapV0.

This also updates the solc version to 0.8.18, the current latest. This
may or may not be required, but there are two solc bugs prior to 0.8.17:
- https://etherscan.io/solcbuginfo?a=StorageWriteRemovalBeforeConditionalTermination
- https://etherscan.io/solcbuginfo?a=AbiReencodingHeadOverflowWithStaticArrayCleanup
Since we are having to re-deploy the contracts, we are updating the
solc version to 0.8.18+commit.87f61d96

(*assetWallet).isRedeemable now performs the redeemable check itself.
That is:
- the swap() contractor method can retrieve the swap state
- the swap state is SSInitiated
- the secret hashes to the secretHash