server/eth: Monitor RPC provider health
Previously, the ETH backend would not start if any of the providers
could not connect or was outdated. Now, if at least one of the
providers is able to connect and the header is recent, the backend
will start. After connecting, a goroutine will start that will
periodically check the health of the RPC providers, sort the list
putting the non-outdated ones first, outdated ones after, and the
ones that fail to respond last. Requests will then attempt to use the
rpc providers in this order.