client/eth: Fix init estimates.
Make estimates aware of the max gas limit for transactions that we later
use when trading. Use one fourth of the block gas limit or the tx fee
limit per max fee per gas as the top end, whichever is lower. Set
simnet's block gas limit to 30 million, the same as test and mainnet.