client/asset/{btc,dcr}: best swap estimate is perfect match
Previously we considered the "best" swap fee estimate to be when there
was a single transaction that batched several matches, but the true best
is when there is a perfect match with another multi-lot order.

The UI and Go interface docs describe this as well.  This makes the
estimates reflect this interpretation.