{client,server}/btc: BTC Fidelity Bonds
This enbales BTC fidelity bonds on both the client and the server.
The implementation is almost identical to DCR. Noteably, on the client,
the `sendToAddress` wallet API is no longer used becuase sends need to
respect the bond reserves. Instead, the transaciton is constructed in
the client and passed to the wallet through `sendRawTransaction`.

An `UnsignedCoinID` field is added to the asset.Bond struct in order to
support BTC wallets that have segwit turned off.