multi: build with eth code by default
This negates the lgpl build tag, making it nolgpl.

The importlgpl.go files in the CLI app (main) packages are updated to
indicate that the `nolgpl` build tag should be used to exclude such
code if it is not acceptble.

Use eth-specific packages now implies lgpl. That is, we gate at the
import of these packages in the cmd apps that would use them via their
importlgpl.go files that require the !nolgpl build tag.

All docs and scripts are updated for the new default.