client/asset/eth: silence provider check spam
This quiets down the logs from the RPC provider compatibility tests,
which was only supposed to be enable for testing and debugging.

This also removes a duplicate check for PendingNonceAt.

This also removes the redundant sublogger so that
CORE[eth][ETH] becomes just CORE[eth].  Same on server.

Finally, the changes the subloger used by CreateWallet not be
CORE[CREATE], and instead just show the asset symbol.