Add ETHBondV0.sol contract
This adds a solidity contract for managing ETH fidelity bonds. In addition
to the usual functionalities of creating and refunding bonds, it also
allows using funds locked in existing bonds to create a new bond. If
a bond with more funds than the existing bonds is being created, the user
has to send additional funds, and if a bond is being created with less
funds than the existing bonds, a "change" bond is also created with the
remainder that maintains the earliest lockTime of the existing bonds. This
process will only allow the user to lock funds for a longer period than
they were previously locked for, never less.

A Hardhat project is also created for this contract to allow easy testing
of the smart contract logic. To run the tests, go to the
`dex/networks/eth/bondcontracts/v0` folder and run the following:
`npm install`
`npx hardhat test`