{client,server}/btc: BTC Fidelity Bonds
This enables BTC fidelity bonds on both the client and the server.
The implementation is almost identical to DCR. Notably, on the client,
the `sendToAddress` wallet API is no longer used because sends need to
respect the bond reserves. Instead, the transaction is constructed in
the client and passed to the wallet through `sendRawTransaction`.