client/asset/eth: Add BondUpdater interface and implement it in ETH Client
Adds a BondUpdater interface which contains a single method, `UpdateBondsTx`.
Similar to MakeBondTx, this generates an unsigned transaction, but
instead of creating a new bond from scratch, it uses the funds locked
in existing bonds to create a new bond.

For ETH, it is not possible to create a signed transaction that will
be submitted later because if the nonce is used before the transaction
is submitted, the transaction will no longer be valid. Therefore, for ETH,
the SignedTx and UnsignedTx in the `asset.Bond` returned from `MakeBondTx`
and `UpdateBondsTx` contain the same data. The ETH wallet backend populates
the nonce and signs the transaction on the call to `SendTransaction`.