client/core: Support wallets that implement 'BondUpdater`
This updates the client to support wallets that implement `BondUpdater`.
In `rotateBonds`, if an account is set to maintain a certain tier, and it
already has some weak or expired bonds posted in an asset that supports
updating bonds, it will use the funds in those bonds to create a new bond
with a sufficiently long lockTime.