server/asset/eth: Add BondUpdater interface and implement in ETH
This adds a BondUpdater interface in the server that requires a backend
to implement an `AllAccountBonds` method. in addition to the Bonder
interface's methods. This method returns all the currenlty active bonds
for an account. This new interface is implemented by the ETH backend. This
is required for assets where it is possible to use existing unexpired
bonds to fund new bonds, because the server needs to know if bonds that it
was previously using the calculates a user's tier no longer exist.