client: Put market making logic outside of core
This moves the market making logic outside of core. There is a new
`MarketMaker` type which calls into core. When running the `MarketMaker`,
one strategy can be defined for each dex market. Since the Gap strategy
was originally designed to have multiple bots running on a market, the
strategy will have to be modified, but this diff only deals with
refactoring the logic outside of core.

The RPCServer is updated with startMarketMaking and stopMarketMaking
commands. Market making can be started by providing a path to a config
file.

New Core methods are also added to support the MarketMaker: `ExchangeMarket`
and `SingleLotFees`.