client/asset/btc: more RPC reconfigure fallbacks to restart
When changing the rpcClient connection config, fall back to full
restart in these added cases:
- when the rpcClient has a custom connectFunc
- when the current deposit address fails to validate with the new
  client, or when it is not owned by the new wallet.

The checks in the second case above were previously only done
if the wallet was actively used.  Now they are always done, but
they are no longer fatal to reconfigure, only live-reconfigure,
causing a full reconnect to do the checks again with a new instance
of an *rpcClient.