client/webserver: more judicious use of authMiddleware
The authMiddleware calls (*Core).User and injects the struct into the
request.  This is primarily needed for html template execution in the
web page handlers.  It is not required for any of the other redirection
middleware or the api endpoints.  And it is definitely not required
for fileServer (img/css/js resources).

This change restricts the user of authMiddleware to the web page
sub-router.  Almost every page requires it, using the commonArgs
helper to prepare the data for template rendering.

This modifies the redirection middlewares (requireInit, requireLogin,
rejectUnauthed, requireDEXConnection, etc). to only use the Core
methods that are actually required, if any.

Finally, this modifies the two api handlers that required authMiddleware
(actuallyLogin and apiUser) so that they do the requests themselves.