site: bond asset choice wording tweaks and settings page fixes
This tweaks the wording on the bond asset selection form and adds the
"whatsabond" information text.

This also fixes a bug where the settings page was using the old registration fee
instead of the "bonds fee buffer".  It also fixes the condition for skipping the
wallet balance wait form created from the settings page.

---------

Co-authored-by: buck54321 <buck54321@gmail.com>