site: add cache buster utilities and test
Add script, bust_caches.sh, to update cache busters. Set cache busters to a hash of the relevant directory concatenated with a hash of the build output in the dist folder. Ensure source side of cache busters are current in run_tests.sh.
The script ci_tests.sh is used in the "Build JS" workflow jobs to check the output components after building the dist folder fresh.