client/{core, btc}: MultiTrade
This adds a `core.MultiTrade` function. It allows callers to fund multiple
trades in one go. It is only implemented for BTC currently, but will be
implemented for other assets later. In order to fund multiple orders in
one go, the BTC wallet first attempts to fund each of the orders with the
existing UTXOs. If this was not possible and splitting is not allowed, the
orders that are able to be funded are returned. If splitting is allowed, a
split transaction with one output per order will be created. As few orders
as possible are included into the split.

This also updates `FundOrder` to return the fees spent on any split
transactions, and this is stored in the database.