client/mm: Update Basic Market Maker
This updates the basic market maker strategy to place orders at multiple
distances from the basis price instead of just one. It uses the new
MultiTrade function to do this.

The basis price calculation is also updated to use fiat rates as a
fallback if there are no trades on the dex or an oracle rate.