testing/loadbot: Add moving markets. (#2409)
Add several tools to the loadbot to allow for sideways, trending, and
volatile markets. Remove most mining as continuous mining is provided by
the harnesses themselves, and this allows us to run multiple bots at
once without excessive mining.

To allow for sideways and trending markets, new options for the
sidestacker can be used. These are a flat increase to the rate or an
automatic oscillating effect. Volatile markets can be achieved by
running the whale bot separately that randomly buys, or sells, large
sums moving the market violently.