client/{webserver,app}: Update Market Making UI
Updates the market making UI to reflect the updated way in which market
making works. A JSON configuration file, the same that would be used by
the RPC server, is stored on disk, and the UI gives users an interactive
way to read/update the config file, and also to start/stop market making.

The updated UI includes two screens, `/mm` and `/mmsettings`. `/mm` allows
users to see each of the market for which a market maker is configured,
allows the user to add/remove market makers, and links to the `/mmsettings`
page of each market maker. The `/mmsettings` page allows users to configure
each setting of each of the market makers.

The markets page is also updated to not allow users to place orders manually
while market making is running.