client/asset/eth: Transaction History
This diff updates the ETH wallet to support the asset.WalletHistorian
interface. Since ETH RPC nodes do not support querying the transactions
an account has made, the transactions are stored locally in a database.
Initially they are stored with a block number of 0 and the max possible
fees, but when the transaction is confirmed, these values are updated.

The pending transaction tracking used for more accurate balance reporting
is updated to use the same data as required for tracking pending
transactions for the wallet history. Also, the database used for monitoring
pending redemption transactions and resubmitting them is upgraded to be
used for both that functionality and the new tx history functionality.