client/mm: Arb Market Maker
This adds a market making strategy that is kind of a combination
between the basic market maker and the simple arbitrage strategies.
Based on a CEX order book, it places orders on the DEX order book, and
when there is a match on the DEX, the opposite order on the DEX is
immediately executed for a profit.