client/asset/btc: SPV Wallet Transaction History
This diff implements `TransactionHistory` for btc SPV wallets. A database
is created to store all the transactions related to the wallet. Although
the SPV wallet stores some transaction information already, the list
transactions output is unreliable and confusing, so it is better to not
parse it every time the tx history is requested. Also, it is not possible
to differentiate a send and a swap transaction.

Incoming transactions are discovered by querying
`ListTransactionsSinceBlock` every block. There is also special handling
for bond creation transactions which are not immediately submitted.