reputation and limits ui
Adapt registration forms to bonding. Show trading limits for the
tier input value. Add bond data to bond options registration form.
Reuse registration forms for updating bond options on dexsettings.
Show tier and reputation on dexsettings. Show parcel limits and
reputation on markets view.

A couple of notes. Nowhere are we going to show the cost of an
individual bond, because that value is misleading. Instead, we'll
show the "bond lock", which is the amount that will actually be
locked for the given tier.

I have removed the simnetfiatrates settings. fiat rates are all
on by default. I have also changed coinpaprika so that it pulls
all coin data, instead of just the ones with wallets. This enables
us to e.g. show USD conversion of the bond lock value for assets
that we don't yet have a wallet for.