ui: arb + market maker ui
Implement arbmm ui. This required a fair amount refactoring on
both the front and back ends. For the back end, the primary
changes are about when and how cexes are loaded. Since we need
to know what markets are available before the bot it started,
we need to call Markets() on cex initialization, before the CEX
is running. This occurs either during startup (MarketMaker is
now a dex.Connector), or when the cex is added (through the new
UpdateCEXConfig method).

In order to facilitate ui testing, I've made the MarketMaker an
interface for use by WebServer (like clientCore), and then
implemented the interface in live_test.go.

On the front end, most effort is focused around mmsettings, and
there was a lot of refactoring done to facilitate the changes
and clean things up.