client/mm: CEX balance tracking and auto-rebalancing (#2568)
* multi: Return TxID from Send and add TransactionConfirmations

Send and Withdraw are updated to return a transaction ID. Also,
a TransactionConfirmations function is added to the wallet interface
which allows callers to determine the amount of confirmations a wallet
transaction has.

* client/mm/libxc: Add deposit/withdraw functionality to CEX interface

- Deposit and Withdraw functionality is added to the CEX interface
and is implemented in Binance.
- The CEX interface is updated to take uint32 assetIDs instead of strings
to represent assets.
- The "weth.polygon" and "wbtc.polygon" assets are renamed to be
"eth.polygon" and "btc.polygon". This makes it clear that "btc" and
"btc.polygon" are the same asset, just on different networks.

* client/mm: Tracking of CEX balances

Tracking of CEX balances for each bot is implemented in a similar fashion
to tracking bot balances on the DEX. A `wrappedCEX` is created for each
bot which behaves as if the entire balance of the CEX is the amount that
is allocated for each bot.

* client/mm: Simple arb bot auto-rebalancing

This implements auto-rebalancing in the simple arbitrage strategy. This is
an optional functionality that allows the user to specify a minimum
balance of each asset that should be on both the DEX and the CEX. If the
balance dips below this amount, then either a deposit or withdrawal will
be done to have an equal amount of the asset on both exchanges.