client/mm/libxc: Binance orderbook diff updates
Previously the Binance library was subscribing to a market data stream from
Binance that sent the 20 orders closest to the mid-gap in every update.
Now, it fetches a snapshot of 1000 orders, this is stored in memory, and
is updated based on a stream that sends diffs instead.
Other minor cleanup is done in the Binance library as well.